Move separation helpers to internals/separation.py
- Relocates check_for_separation (renamed from _check_for_separation), the
  _SeparationMethod Protocol, and the _fe / _ir variants from
  models/fepois_.py to pyfixest/estimation/internals/separation.py.
- Updates the two call sites (feglm_.py, fepois_.py) to import from the new
  location.
- Drops now-unused imports in fepois_.py (re, importlib.import_module,
  typing.Protocol).

Pure relocation: no behavior change. Also unblocks the upcoming Fepois→Feglm
fold by removing the awkward cross-import (feglm_ → fepois_ for a helper that
isn't really Poisson-specific).
